expense with user authentication,and deleteExpense

+++ controllers/expense.js
@@ -1,4 +1,5 @@
 const Expense=require('../models/expenses');
+const jwt = require('jsonwebtoken');
 
 function isStringinvalid(string){
   if(string ==undefined || string.length== 0){
@@ -17,7 +18,9 @@ exports.addExpenses= async(req,res,next)=>{
         return res.status(400).json({err:'invalid input:Something is missing'})
     }
 
-    const data= await Expense.create({Expenseamount:Expenseamount,description:description,catagory:catagory});
+   //can be written: const data =await req.createExpense({Expenseamount:Expenseamount,description:description,catagory:catagory})
+
+    const data= await Expense.create({Expenseamount:Expenseamount,description:description,catagory:catagory,userId:req.user.id});
       res.status(201).json({expenseDetails:data,success:true,message:'Expense added'});
     }catch(err){
       res.status(500).json({error:err})
@@ -26,10 +29,12 @@ exports.addExpenses= async(req,res,next)=>{
 
 exports.getExpenses = async(req,res,next)=>{
     try{
-    const expenses = await Expense.findAll();
+    const expenses = await Expense.findAll({where:{userId:req.user.id}});
+
     console.log(expenses);
     res.status(200).json({expenseDetails:expenses});
   }catch(err){
+    console.log(err)
     res.status(500).json({error:err});
   }
   }
@@ -42,9 +47,17 @@ exports.getExpenses = async(req,res,next)=>{
   
     }
     const expenseId=req.params.id;
-    await Expense.destroy({where:{id:expenseId}});
-     res.sendStatus(200);
+    await Expense.destroy({where:{id:expenseId, userId:req.user.id}}).then((numRows)=>{
+      if(numRows==0){
+        return res.status(404).json({success:false,message:'expense not belongs to User'})
+      }
+     return res.sendStatus(200).json({success:true,message:"deleted successfully"});
+    }).catch(err=>{
+      console.log(err)
+    })
+    
    }catch(err){
+    console.log(err)
       res.status(500).json({error:err});
       
     }
